change: re-organize schema module
Re-organize schema module to allow its functions as much as possible
even if the required module is not available.

- src/anyconfig/schema: re-organize the module with some refactoring
- tests/schema: likewise
- tests/api, tests/cli: follow changes in anyconfig.schema
