+++ src/anyconfig/schema/__init__.py
@@ -2,15 +2,29 @@
 # Copyright (C) 2021 - 2024 Satoru SATOH <satoru.satoh gmail.com>
 # SPDX-License-Identifier: MIT
 #
-"""Misc global constants, variables, classes and so on."""
+"""Schema generation and validation."""
+from __future__ import annotations
+
+from .jsonschema.generator import gen_schema
+
 try:
-    from .jsonschema import validate, is_valid, gen_schema
+    from .jsonschema.validator import validate, is_valid
+
+    VALIDATORS = {
+        "jsonschema": validate
+    }
     SUPPORTED: bool = True
 except ImportError:
-    from .default import validate, is_valid, gen_schema
+    from .default import validate, is_valid  # noqa: F401
+    VALIDATORS = {}
     SUPPORTED = False
 
 
-__all__ = [
-    "validate", "is_valid", "gen_schema", "SUPPORTED"
+GENERATORS = {
+    "jsonschema": gen_schema
+}
+
+_all__ = [
+    "validate", "is_valid", "gen_schema",
+    "VALIDATORS", "GENERATORS", "SUPPORTED"
 ]
