+++ src/anyconfig/schema/datatypes.py
@@ -2,10 +2,15 @@
 # Copyright (C) 2021 - 2024 Satoru SATOH <satoru.satoh gmail.com>
 # SPDX-License-Identifier: MIT
 #
+# pylint: disable=unused-import
 """Some common constants, utility functions and so on."""
 from __future__ import annotations
 
 import typing
+
+from ..common import (  # noqa: F401
+    ValidationError, InDataT, InDataExT
+)
 
 
 ResultT = tuple[bool, typing.Union[str, list[str]]]
