+++ src/anyconfig/schema/jsonschema/__init__.py
@@ -1,0 +1,1 @@
+"""jsonschema generator and validator."""
